[SEDONA-100] Add st_multi function
